requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
pypdf
pydantic